- The kernel now saves its errno on process switch(maybe save it on every thread switch?); - Implemented major and minor device numbers.

+++ libc/include/sys/types.h
@@ -79,7 +79,7 @@ typedef unsigned long id_t;
 
 #ifndef __dev_t_defined
 #define __dev_t_defined
-typedef long dev_t;
+typedef unsigned int dev_t;
 #endif
 
 #endif
